Merge tag 'jdk-25+26' into lworld_merge_jdk_25_26

Added tag jdk-25+26 for changeset c59e44a7

+++ src/java.base/share/classes/java/time/YearMonth.java
@@ -161,7 +161,7 @@ public final class YearMonth
     /**
      * @serial The month-of-year, not null.
      */
-    private final int month;
+    private final byte month;
 
     //-----------------------------------------------------------------------
     /**
@@ -314,7 +314,7 @@ public final class YearMonth
      */
     private YearMonth(int year, int month) {
         this.year = year;
-        this.month = month;
+        this.month = (byte) month;
     }
 
     /**
